+++ randovania/exporter/hints/pickup_hint.py
@@ -8,7 +8,7 @@ from randovania.exporter.hints.determiner import Determiner
 from randovania.game.game_enum import RandovaniaGame
 from randovania.game_description.assignment import PickupAssignment, PickupTarget
 from randovania.game_description.hint import HintItemPrecision
-from randovania.game_description.hint_features import HintDetails, PickupHintFeature
+from randovania.game_description.hint_features import HintDetails, HintFeature
 from randovania.game_description.resources.pickup_index import PickupIndex
 from randovania.generator.pickup_pool import pickup_creator
 
@@ -67,7 +67,7 @@ def _calculate_determiner(pickup_assignment: PickupAssignment, pickup: PickupEnt
 def create_pickup_hint(
     pickup_assignment: PickupAssignment,
     region_list: RegionList,
-    precision: HintItemPrecision | PickupHintFeature,
+    precision: HintItemPrecision | HintFeature,
     target: PickupTarget | None,
     players_config: PlayersConfiguration,
     include_owner: bool,
@@ -91,12 +91,8 @@ def create_pickup_hint(
             player=players_config.player_index,
         )
 
-    if isinstance(precision, PickupHintFeature):
+    if isinstance(precision, HintFeature):
         details = precision.hint_details
-
-    elif precision is HintItemPrecision.BROAD_CATEGORY:
-        broad = next(category for category in target.pickup.hint_features if category.is_broad_category)
-        details = broad.hint_details
 
     elif precision is HintItemPrecision.DETAILED:
         details = HintDetails(_calculate_determiner(pickup_assignment, target.pickup, region_list), target.pickup.name)