+++ randovania/game_description/hint.py
@@ -1,18 +1,21 @@
 from __future__ import annotations
 
+import dataclasses
 import typing
 from dataclasses import dataclass
 from enum import Enum
 
 from randovania.bitpacking.json_dataclass import JsonDataclass
+from randovania.bitpacking.type_enforcement import DataclassPostInitTypeCheck
 from randovania.game_description.db.area_identifier import AreaIdentifier
-from randovania.game_description.hint_features import HintFeature, PickupHintFeature
+from randovania.game_description.hint_features import HintFeature
 from randovania.game_description.resources.pickup_index import PickupIndex
 from randovania.lib import enum_lib
 
 if typing.TYPE_CHECKING:
     from randovania.game_description.game_description import GameDescription
     from randovania.game_description.pickup.pickup_database import PickupDatabase
+    from randovania.generator.hint_distributor import HintFeatureGaussianParams
 
 
 class HintDarkTemple(Enum):
@@ -31,10 +34,6 @@ class HintItemPrecision(Enum):
     # The exact item
     DETAILED = "detailed"
 
-    # x-related, life-support, or just the precise category
-    # TODO: remove this if keybearer hints are rebalanced
-    BROAD_CATEGORY = "broad-category"
-
 
 class HintLocationPrecision(Enum):
     # Precision hasn't been assigned yet
@@ -52,6 +51,21 @@ class HintLocationPrecision(Enum):
     # DEPRECATED: Relative hints
     RELATIVE_TO_AREA = "relative-to-area"
     RELATIVE_TO_INDEX = "relative-to-index"
+
+
+@dataclass(frozen=True, order=True)
+class SpecificHintPrecision(DataclassPostInitTypeCheck):
+    """
+    Can be used in a PrecisionPair to later be replaced with
+    a HintFeature close to this exact precision.
+    """
+
+    mean: float = dataclasses.field(metadata={"min": 0.0, "max": 1.0})
+    std_dev: float = dataclasses.field(default=0.0, metadata={"min": 0.0})
+
+    @property
+    def gauss_params(self) -> HintFeatureGaussianParams:
+        return self.mean, self.std_dev
 
 
 class HintRelativeAreaName(Enum):
@@ -77,14 +91,14 @@ class RelativeData:
 @dataclass(frozen=True)
 class RelativeDataItem(JsonDataclass, RelativeData):
     other_index: PickupIndex
-    precision: HintItemPrecision | PickupHintFeature
+    precision: HintItemPrecision | HintFeature
 
     @classmethod
     def from_json(cls, json_dict: dict, **extra: typing.Any) -> typing.Self:
         pickup_database: PickupDatabase = extra["other_pickup_db"]
 
         item_json = json_dict.get("precision")
-        item: HintItemPrecision | PickupHintFeature
+        item: HintItemPrecision | HintFeature
         if item_json is not None:
             item = HintItemPrecision(item_json)
         else:
@@ -99,7 +113,7 @@ class RelativeDataItem(JsonDataclass, RelativeData):
     @property
     def as_json(self) -> dict:
         data = super().as_json
-        if isinstance(self.precision, PickupHintFeature):
+        if isinstance(self.precision, HintFeature):
             del data["precision"]
             data["precision_feature"] = self.precision.name
         return data
@@ -113,8 +127,8 @@ class RelativeDataArea(JsonDataclass, RelativeData):
 
 @dataclass(frozen=True)
 class PrecisionPair(JsonDataclass):
-    location: HintLocationPrecision | HintFeature
-    item: HintItemPrecision | PickupHintFeature
+    location: HintLocationPrecision | HintFeature | SpecificHintPrecision
+    item: HintItemPrecision | HintFeature | SpecificHintPrecision
     include_owner: bool | None = None
     relative: RelativeDataItem | RelativeDataArea | None = None
 
@@ -133,7 +147,7 @@ class PrecisionPair(JsonDataclass):
             location = game.hint_feature_database[json_dict["location_feature"]]
 
         item_json = json_dict.get("item")
-        item: HintItemPrecision | PickupHintFeature
+        item: HintItemPrecision | HintFeature
         if item_json is not None:
             item = HintItemPrecision(item_json)
         else:
@@ -148,11 +162,14 @@ class PrecisionPair(JsonDataclass):
 
     @property
     def as_json(self) -> dict:
+        if isinstance(self.location, SpecificHintPrecision) or isinstance(self.item, SpecificHintPrecision):
+            raise TypeError(f"PrecisionPair {self} with SpecificHintPrecision is not JSON encodable!")
+
         data = super().as_json
         if isinstance(self.location, HintFeature):
             del data["location"]
             data["location_feature"] = self.location.name
-        if isinstance(self.item, PickupHintFeature):
+        if isinstance(self.item, HintFeature):
             del data["item"]
             data["item_feature"] = self.item.name
         return data
@@ -219,6 +236,8 @@ def is_unassigned_location(hint: BaseHint) -> typing.TypeGuard[LocationHint]:
         (hint.precision is PRECISION_PAIR_UNASSIGNED)
         or (hint.precision.location is HintLocationPrecision.FEATURAL)
         or (hint.precision.item is HintItemPrecision.FEATURAL)
+        or isinstance(hint.precision.location, SpecificHintPrecision)
+        or isinstance(hint.precision.item, SpecificHintPrecision)
     )