+++ randovania/games/prime2/generator/hint_distributor.py
@@ -4,12 +4,14 @@ from typing import TYPE_CHECKING, override
 
 from randovania.game.game_enum import RandovaniaGame
 from randovania.game_description import default_database
+from randovania.game_description.db.hint_node import HintNodeKind
 from randovania.game_description.db.node_identifier import NodeIdentifier
 from randovania.game_description.hint import (
     HintDarkTemple,
     HintItemPrecision,
     PrecisionPair,
     RedTempleHint,
+    SpecificHintPrecision,
 )
 from randovania.game_description.resources.pickup_index import PickupIndex
 from randovania.games.prime2.layout.echoes_configuration import EchoesConfiguration
@@ -17,12 +19,12 @@ from randovania.generator.hint_distributor import HintDistributor, HintTargetPre
 from randovania.lib import enum_lib
 
 if TYPE_CHECKING:
+    from collections.abc import Container
     from random import Random
 
     from randovania.game_description.game_patches import GamePatches
     from randovania.game_description.pickup.pickup_entry import PickupEntry
     from randovania.generator.filler.filler_configuration import PlayerPool
-    from randovania.generator.filler.player_state import PlayerState
     from randovania.generator.pre_fill_params import PreFillParams
 
 
@@ -80,12 +82,12 @@ class EchoesHintDistributor(HintDistributor):
         patches: GamePatches,
         rng: Random,
         player_pool: PlayerPool,
-        player_state: PlayerState,
         player_pools: list[PlayerPool],
+        hint_kinds: Container[HintNodeKind] = {HintNodeKind.GENERIC},
     ) -> GamePatches:
         assert isinstance(player_pool.configuration, EchoesConfiguration)
         if player_pool.configuration.hints.item_hints:
-            return self.add_hints_precision(player_state, patches, rng, player_pools)
+            return await super().assign_precision_to_hints(patches, rng, player_pool, player_pools, hint_kinds)
         else:
             return self.replace_hints_without_precision_with_jokes(patches)
 
@@ -94,7 +96,7 @@ class EchoesHintDistributor(HintDistributor):
         game = default_database.game_description_for(RandovaniaGame.METROID_PRIME_ECHOES)
         keybearer = game.hint_feature_database["specific_hint_keybearer"]
 
-        precision = PrecisionPair(keybearer, HintItemPrecision.BROAD_CATEGORY, include_owner=True)
+        precision = PrecisionPair(keybearer, SpecificHintPrecision(0.4), include_owner=True)
 
         c = NodeIdentifier.create
         keybearers = [