+++ test/games/prime/patcher_file_lib/test_item_hints.py
@@ -280,7 +280,6 @@ def test_create_message_for_hint_dark_temple_no_keys(empty_patches, players_conf
     "item",
     [
         (HintItemPrecision.DETAILED, "The", "&push;&main-color=#FF6705B3;Blank Pickup&pop;"),
-        (HintItemPrecision.BROAD_CATEGORY, "A", "&push;&main-color=#FF6705B3;life support system&pop;"),
     ],
 )
 @pytest.mark.parametrize(
@@ -343,7 +342,6 @@ def test_create_hints_item_location(echoes_game_patches, blank_pickup, item, loc
     "item",
     [
         (HintItemPrecision.DETAILED, "the &push;&main-color=#FF6705B3;Blank Pickup&pop;"),
-        (HintItemPrecision.BROAD_CATEGORY, "a &push;&main-color=#FF6705B3;life support system&pop;"),
     ],
 )
 def test_create_hints_guardians(
@@ -390,7 +388,6 @@ def test_create_hints_guardians(
     "item",
     [
         (HintItemPrecision.DETAILED, "the &push;&main-color=#FF6705B3;Blank Pickup&pop;"),
-        (HintItemPrecision.BROAD_CATEGORY, "a &push;&main-color=#FF6705B3;life support system&pop;"),
     ],
 )
 def test_create_hints_light_suit_location(echoes_game_patches, players_config, blank_pickup, item):
@@ -429,7 +426,6 @@ def test_create_hints_light_suit_location(echoes_game_patches, players_config, b
     ("reference_precision", "reference_name"),
     [
         (HintItemPrecision.DETAILED, "the Reference Pickup"),
-        (HintItemPrecision.BROAD_CATEGORY, "a life support system"),
     ],
 )
 @pytest.mark.parametrize(