+++ test/generator/pickup_pool/test_pickup_creator.py
@@ -5,7 +5,7 @@ import dataclasses
 import pytest
 from frozendict import frozendict
 
-from randovania.game_description.hint_features import HintDetails, PickupHintFeature
+from randovania.game_description.hint_features import HintDetails, HintFeature
 from randovania.game_description.pickup.pickup_definition.ammo_pickup import AmmoPickupDefinition
 from randovania.game_description.pickup.pickup_definition.standard_pickup import StandardPickupDefinition
 from randovania.game_description.pickup.pickup_entry import (
@@ -22,8 +22,8 @@ from randovania.layout.base.standard_pickup_state import StandardPickupState
 
 
 @pytest.fixture
-def less_generic_pickup_category() -> PickupHintFeature:
-    return PickupHintFeature(
+def less_generic_pickup_category() -> HintFeature:
+    return HintFeature(
         name="the_category",
         long_name="The Category",
         hint_details=HintDetails("a ", " wonderful item"),
@@ -176,7 +176,6 @@ def test_create_missile_launcher(
                 categories["missile"],
                 categories["missile_related"],
                 categories["chozo"],
-                categories["weapon"],
             )
         ),
         generator_params=default_generator_params,
@@ -239,7 +238,6 @@ def test_create_seeker_launcher(
                 categories["missile_related"],
                 categories["major"],
                 categories["luminoth"],
-                categories["weapon"],
             )
         ),
         respects_lock=ammo_requires_main_item,